Set repo in manifest to point to rust-bitcoin org
We just transferred this repo to the `rust-bitcoin` org. We need to
update the link on `crates.io` because I remove the auto-redirect from
my github account so I can fork the repo to work on it. To do so we need
to release a new version, can be done as a patch version.

Update the manifest, bump the patch version, update the lock-files.